delete dead code and rename pure->web100
